pyproject.toml
[tool.poetry]
name = "fastapi-scalable"
version = "0.1.0"
description = ""
authors = ["Joje1283 <joje1283@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
passlib = "^1.7.4"
pydantic-settings = "^2.6.1"
python-jose = "^3.3.0"
sqlmodel = "^0.0.22"
dependency-injector = "^4.43.0"
loguru = "^0.7.2"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
celery = "^5.4.0"
py-ulid = "^1.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
